# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SwipeRx Operator",
    "version": "Alpha 0.1",
    "description": "Ninja Van Indonesia's delivery-compliance and returns tool for the SwipeRx pharmaceutical account. Implant/DE staff turn a SwipeRx TMP batch file into a Ninja Van order-creation upload with a tokenized courier link per AWB; couriers open that link at the pharmacy to capture the required POD/RDO photos and flag rejected goods; ops then acknowledge each reject and record the replacement return TIDs.\n\nStaff endpoints are authenticated by the Substrait SSO gateway (X-Forwarded-Email) and authorized by role. The /api/c/{token} courier endpoints are deliberately unauthenticated: the link token is itself the credential, because riders open them from a delivery instruction on a phone with no account."
  },
  "servers": [
    {
      "url": "https://swiperx-operator.ninjavan.apps.substrait.build",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/api/returns": {
      "get": {
        "operationId": "list_returns",
        "tags": [
          "returns"
        ],
        "summary": "Returns up to 500 reject-return rows newest-first with their derived stage, closing pipeline, origin-unknown flag and at-the-door proof photos, optionally filtered to one derived stage.",
        "parameters": [
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "description": "Filter on the DERIVED stage. Rejected with 400 bad_stage if not one of the listed values.",
            "schema": {
              "type": "string",
              "enum": [
                "pending_de_upload",
                "pending_print",
                "printed",
                "rts_triggered",
                "tids_sent"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The worklist plus the RTS shipper account the deck mandates.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "returns",
                    "rts_shipper_id"
                  ],
                  "properties": {
                    "returns": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ReturnRow"
                      }
                    },
                    "rts_shipper_id": {
                      "type": "string",
                      "description": "Constant RTS_SHIPPER_ID.",
                      "example": "11398434"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "bad_stage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "not_authenticated / invalid_session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "403": {
            "description": "forbidden — requires implant, de, station_ic or program_manager (superadmin always passes)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/returns/export.csv": {
      "get": {
        "operationId": "export_csv",
        "tags": [
          "returns"
        ],
        "summary": "Downloads the whole worklist flattened to one audit CSV row per return, carrying every stage timestamp and the email of whoever stamped it, including the legacy validator and pasted-TID columns.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "UTF-8 BOM + CSV, columns: return_id, forward_awb, return_awb, pharmacy, city, reject_type, rejected_at, stage, closes_by, hub, origin, reject_pcs, legacy_validated_at, legacy_validated_by, de_uploaded_at, de_uploaded_by, printed_at, printed_by, rts_requested_at, rts_requested_by, flagged_at, flagged_by, flag_note, legacy_return_tids.",
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "description": "attachment; filename=\"reject-returns.csv\""
              }
            },
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "not_authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "403": {
            "description": "forbidden — implant, de, station_ic, program_manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/vasil/returns": {
      "get": {
        "operationId": "vasil_list_returns",
        "tags": [
          "vasil"
        ],
        "summary": "Lists every reject return (all stages) for the Migrate to Vasil Operator page, filtered to the pending, done or all tab; rows hidden from the page never appear. Superadmin and KAM only.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "done",
                "all"
              ]
            },
            "description": "Which tab. Anything else is 400 bad_status."
          }
        ],
        "responses": {
          "200": {
            "description": "Rows in the same shape as GET /api/returns, each with vasil_status (pending|done), vasil_done_at and vasil_done_by_email.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "returns": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/vasil/export.csv": {
      "get": {
        "operationId": "vasil_export_csv",
        "tags": [
          "vasil"
        ],
        "summary": "Downloads the Vasil page's pending, done or all tab as a CSV with exactly the same columns as the Reject returns export. Superadmin and KAM only.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "done",
                "all"
              ]
            },
            "description": "Which tab. Anything else is 400 bad_status."
          }
        ],
        "responses": {
          "200": {
            "description": "UTF-8 BOM + CSV; filename vasil-<status>.csv.",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/vasil/mark-done": {
      "post": {
        "operationId": "vasil_mark_done",
        "tags": [
          "vasil"
        ],
        "summary": "Ticks the selected pending AWBs as uploaded to Vasil Operator; rows already done or hidden are skipped. Touches only vasil_migration, never return_parcel.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ids"
                ],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int64"
                    },
                    "description": "return_parcel.id values; empty is 400 no_ids."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Count of rows moved to done.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "updated": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/vasil/mark-pending": {
      "post": {
        "operationId": "vasil_mark_pending",
        "tags": [
          "vasil"
        ],
        "summary": "Moves the selected done AWBs back to pending (undo for a mis-click); clears who/when it was marked.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ids"
                ],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int64"
                    },
                    "description": "return_parcel.id values; empty is 400 no_ids."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Count of rows moved back to pending.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "updated": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/vasil/hide": {
      "post": {
        "operationId": "vasil_hide",
        "tags": [
          "vasil"
        ],
        "summary": "Removes the selected AWBs from the Vasil page only (status hidden); the reject return itself is never deleted and stays on Reject returns.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ids"
                ],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int64"
                    },
                    "description": "return_parcel.id values; empty is 400 no_ids."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Count of rows hidden.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "updated": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPError": {
        "type": "object",
        "description": "FastAPI's HTTPException body. 'detail' is a short machine-readable slug (e.g. not_authenticated, forbidden, not_found, bad_stage) or, on 422 from the OC engine, a free-text parse message.",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
...
      "ProofPhoto": {
        "type": "object",
        "description": "At-the-door evidence attached to a return row (doc types rejected_goods, delivery_note, awb_sticker).",
        "required": [
          "doc_type",
          "photo_url"
        ],
        "properties": {
          "doc_type": {
            "type": "string",
            "enum": [
              "rejected_goods",
              "delivery_note",
              "awb_sticker"
            ]
          },
          "photo_url": {
            "type": "string",
            "description": "Session-gated staff path /api/media/{photo_ref}."
          }
        }
      },
      "ReturnRow": {
        "type": "object",
        "description": "One reject-return worklist row: return_parcel joined to its forward awb and to users for the actor emails, then shaped by _shape(). All timestamps are stringified with str(datetime) ('YYYY-MM-DD HH:MM:SS'), not ISO-8601 'T' format, and are null when unset.",
        "required": [
          "id",
          "original_awb_id",
          "return_type",
          "stage",
          "origin_unknown",
          "closes_by"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "return_parcel.id"
          },
          "original_awb_id": {
            "type": "string",
            "maxLength": 32,
            "description": "Forward tracking number."
          },
          "return_type": {
            "type": "string",
            "enum": [
              "sebagian",
              "semua"
            ],
            "description": "Column default is 'sebagian'; 'semua' means a full refusal."
          },
          "service_id": {
            "type": "string",
            "maxLength": 20,
            "nullable": true
          },
          "return_awb_id": {
            "type": "string",
            "maxLength": 40,
            "nullable": true,
            "description": "The <SwipeAWB>-R01 number, stamped only at mark-uploaded for partials."
          },
          "rejected_at": {
            "type": "string",
            "nullable": true,
            "description": "return_parcel.created_at."
          },
          "acknowledged_at": {
            "type": "string",
            "nullable": true,
            "description": "Legacy pre-24-Aug flow."
          },
          "return_tids": {
            "type": "string",
            "nullable": true,
            "description": "Legacy operator-pasted TIDs, comma/newline separated."
          },
          "tids_sent_at": {
            "type": "string",
            "nullable": true
          },
          "rts_requested_at": {
            "type": "string",
            "nullable": true
          },
          "reject_pcs": {
            "type": "integer",
            "nullable": true,
            "description": "Pieces the driver counted back at the door."
          },
          "validated_at": {
            "type": "string",
            "nullable": true,
            "description": "Legacy validator stamp; no longer gates anything."
          },
          "de_uploaded_at": {
            "type": "string",
            "nullable": true
          },
          "printed_at": {
            "type": "string",
            "nullable": true
          },
          "origin": {
            "type": "string",
            "nullable": true,
            "enum": [
              "TMP_DEPOK",
              "TMP_SURABAYA",
              null
            ],
            "description": "COALESCE(return_parcel.origin, awb.origin)."
          },
          "pharmacy_name": {
            "type": "string",
            "maxLength": 200,
            "nullable": true
          },
          "city": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          },
          "hub_name": {
            "type": "string",
            "maxLength": 32,
            "nullable": true,
            "description": "From the forward awb, set by the driver at identity."
          },
          "phone": {
            "type": "string",
            "maxLength": 40,
            "nullable": true
          },
          "address": {
            "type": "string",
            "maxLength": 400,
            "nullable": true
          },
          "validated_by_email": {
            "type": "string",
            "nullable": true
          },
          "de_uploaded_by_email": {
            "type": "string",
            "nullable": true
          },
          "printed_by_email": {
            "type": "string",
            "nullable": true
          },
          "rts_requested_by_email": {
            "type": "string",
            "nullable": true
          },
          "stage": {
            "type": "string",
            "enum": [
              "pending_de_upload",
              "pending_print",
              "printed",
              "rts_triggered",
              "tids_sent"
            ],
            "description": "Derived from the timestamps above, never stored."
          },
          "origin_unknown": {
            "type": "boolean",
            "description": "True when no origin is recorded; blocks export and mark-uploaded."
          },
          "po_number": {
            "type": "string",
            "nullable": true,
            "description": "The PO the courier said the returned goods came from. The return OC's requested_tracking_number is this + '1'. Falls back to the forward order's only PO line where it has exactly one."
          },
          "po_options": {
            "type": "array",
            "description": "Only populated on po_unknown rows: that forward order's own po_line entries, the same list the courier saw — what a superadmin may choose from via POST /api/returns/po.",
            "items": {
              "type": "object",
              "properties": {
                "po_number": {
                  "type": "string"
                },
                "koli": {
                  "type": "integer"
                }
              }
            }
          },
          "po_unknown": {
            "type": "boolean",
            "description": "A partial reject with no PO — filed before couriers were asked (8 Sep 2026). Blocks export and mark-uploaded, and unlike origin there is no bulk-set: only the courier at the counter could have answered. Always false for a full refusal, which produces no OC row."
          },
          "closes_by": {
            "type": "string",
            "enum": [
              "rts",
              "return_oc"
            ],
            "description": "'rts' for semua, 'return_oc' for sebagian."
          },
          "flagged": {
            "type": "boolean",
            "description": "A Station IC remark is hanging on this row. Never a stage — it moves nothing."
          },
          "flag_note": {
            "type": "string",
            "nullable": true,
            "description": "The current remark, e.g. 'AWB tidak ditemukan di NV / OPV2'."
          },
          "flagged_at": {
            "type": "string",
            "nullable": true
          },
          "flagged_by_email": {
            "type": "string",
            "nullable": true
          },
          "proof_photos": {
            "type": "array",
            "description": "Only populated by GET /api/returns; absent from the internal row shape used by the write endpoints.",
            "items": {
              "$ref": "#/components/schemas/ProofPhoto"
            }
          }
        }
      },
...
    }
  }
}